svg/painting/parsing/shape-rendering-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing shape-rendering with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ShapeRenderingProperty" rel="help"/><h:meta content="shape-rendering supports the full grammar 'auto | optimizeSpeed | crispEdges | geometricPrecision'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("shape-rendering", "auto");
test_valid_value("shape-rendering", "optimizespeed");
test_valid_value("shape-rendering", "crispedges");
test_valid_value("shape-rendering", "geometricprecision");

  ]]></script></svg>
