Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/shape-rendering-valid-expected.txt


PASS e.style['shape-rendering'] = "auto" should set the property value
FAIL e.style['shape-rendering'] = "optimizespeed" should set the property value assert_equals: serialization should be canonical expected "optimizespeed" but got "optimizeSpeed"
PASS e.style['shape-rendering'] = "crispedges" should set the property value
FAIL e.style['shape-rendering'] = "geometricprecision" should set the property value assert_equals: serialization should be canonical expected "geometricprecision" but got "geometricPrecision"
